+++ scripts/moduleConfigs/databaseModules/supabase/nuxtConfig.ts
@@ -1,0 +1,30 @@
+export default {
+  supabase: {
+    url: process.env.NUXT_PUBLIC_SUPABASE_URL,
+    key: process.env.NUXT_PUBLIC_SUPABASE_KEY,
+    serviceKey: process.env.NUXT_SUPABASE_SERVICE_KEY,
+    redirect: true,
+    redirectOptions: {
+      login: '/login',
+      callback: '/confirm',
+      include: undefined,
+      exclude: [],
+      cookieRedirect: true,
+    },
+    clientOptions: {
+      auth: {
+        flowType: 'implicit',
+        autoRefreshToken: true,
+        detectSessionInUrl: true,
+        persistSession: true,
+      },
+    },
+    cookieName: 'sb',
+    cookieOptions: {
+      maxAge: 60 * 60 * 8,
+      sameSite: 'lax',
+      secure: true,
+    },
+    types: './types/database.types.ts',
+  }
+}